Move some creation functions to da.creation
Trying to reduce the file size of da.core.
