+++ dask/array/creation.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
-from functools import partial
+from functools import partial, wraps
 import itertools
 
 import numpy as np
@@ -9,6 +9,7 @@ from .core import Array, normalize_chunks, stack
 from .wrap import empty
 from . import chunk
 from ..base import tokenize
+from .. import sharedict
 
 
 def linspace(start, stop, num=50, chunks=None, dtype=None):
@@ -196,3 +197,194 @@ def indices(dimensions, dtype=int, chunks=None):
         )
 
     return grid
+
+
+def eye(N, chunks, M=None, k=0, dtype=float):
+    """
+    Return a 2-D Array with ones on the diagonal and zeros elsewhere.
+
+    Parameters
+    ----------
+    N : int
+      Number of rows in the output.
+    chunks: int
+        chunk size of resulting blocks
+    M : int, optional
+      Number of columns in the output. If None, defaults to `N`.
+    k : int, optional
+      Index of the diagonal: 0 (the default) refers to the main diagonal,
+      a positive value refers to an upper diagonal, and a negative value
+      to a lower diagonal.
+    dtype : data-type, optional
+      Data-type of the returned array.
+
+    Returns
+    -------
+    I : Array of shape (N,M)
+      An array where all elements are equal to zero, except for the `k`-th
+      diagonal, whose values are equal to one.
+    """
+    if not isinstance(chunks, int):
+        raise ValueError('chunks must be an int')
+
+    token = tokenize(N, chunk, M, k, dtype)
+    name_eye = 'eye-' + token
+
+    eye = {}
+    if M is None:
+        M = N
+
+    vchunks = [chunks] * (N // chunks)
+    if N % chunks != 0:
+        vchunks.append(N % chunks)
+    hchunks = [chunks] * (M // chunks)
+    if M % chunks != 0:
+        hchunks.append(M % chunks)
+
+    for i, vchunk in enumerate(vchunks):
+        for j, hchunk in enumerate(hchunks):
+            if (j - i - 1) * chunks <= k <= (j - i + 1) * chunks:
+                eye[name_eye, i, j] = (np.eye, vchunk, hchunk, k - (j - i) * chunks, dtype)
+            else:
+                eye[name_eye, i, j] = (np.zeros, (vchunk, hchunk), dtype)
+    return Array(eye, name_eye, shape=(N, M),
+                 chunks=(chunks, chunks), dtype=dtype)
+
+
+@wraps(np.diag)
+def diag(v):
+    name = 'diag-' + tokenize(v)
+    if isinstance(v, np.ndarray):
+        if v.ndim == 1:
+            chunks = ((v.shape[0],), (v.shape[0],))
+            dsk = {(name, 0, 0): (np.diag, v)}
+        elif v.ndim == 2:
+            chunks = ((min(v.shape),),)
+            dsk = {(name, 0): (np.diag, v)}
+        else:
+            raise ValueError("Array must be 1d or 2d only")
+        return Array(dsk, name, chunks, dtype=v.dtype)
+    if not isinstance(v, Array):
+        raise TypeError("v must be a dask array or numpy array, "
+                        "got {0}".format(type(v)))
+    if v.ndim != 1:
+        if v.chunks[0] == v.chunks[1]:
+            dsk = dict(((name, i), (np.diag, row[i])) for (i, row)
+                       in enumerate(v._keys()))
+            return Array(sharedict.merge(v.dask, (name, dsk)), name, (v.chunks[0],), dtype=v.dtype)
+        else:
+            raise NotImplementedError("Extracting diagonals from non-square "
+                                      "chunked arrays")
+    chunks_1d = v.chunks[0]
+    blocks = v._keys()
+    dsk = {}
+    for i, m in enumerate(chunks_1d):
+        for j, n in enumerate(chunks_1d):
+            key = (name, i, j)
+            if i == j:
+                dsk[key] = (np.diag, blocks[i])
+            else:
+                dsk[key] = (np.zeros, (m, n))
+
+    return Array(sharedict.merge(v.dask, (name, dsk)), name, (chunks_1d, chunks_1d),
+                 dtype=v.dtype)
+
+
+def triu(m, k=0):
+    """
+    Upper triangle of an array with elements above the `k`-th diagonal zeroed.
+
+    Parameters
+    ----------
+    m : array_like, shape (M, N)
+        Input array.
+    k : int, optional
+        Diagonal above which to zero elements.  `k = 0` (the default) is the
+        main diagonal, `k < 0` is below it and `k > 0` is above.
+
+    Returns
+    -------
+    triu : ndarray, shape (M, N)
+        Upper triangle of `m`, of same shape and data-type as `m`.
+
+    See Also
+    --------
+    tril : lower triangle of an array
+    """
+    if m.ndim != 2:
+        raise ValueError('input must be 2 dimensional')
+    if m.shape[0] != m.shape[1]:
+        raise NotImplementedError('input must be a square matrix')
+    if m.chunks[0][0] != m.chunks[1][0]:
+        msg = ('chunks must be a square. '
+               'Use .rechunk method to change the size of chunks.')
+        raise NotImplementedError(msg)
+
+    rdim = len(m.chunks[0])
+    hdim = len(m.chunks[1])
+    chunk = m.chunks[0][0]
+
+    token = tokenize(m, k)
+    name = 'triu-' + token
+
+    dsk = {}
+    for i in range(rdim):
+        for j in range(hdim):
+            if chunk * (j - i + 1) < k:
+                dsk[(name, i, j)] = (np.zeros, (m.chunks[0][i], m.chunks[1][j]))
+            elif chunk * (j - i - 1) < k <= chunk * (j - i + 1):
+                dsk[(name, i, j)] = (np.triu, (m.name, i, j), k - (chunk * (j - i)))
+            else:
+                dsk[(name, i, j)] = (m.name, i, j)
+    return Array(sharedict.merge((name, dsk), m.dask), name,
+                 shape=m.shape, chunks=m.chunks, dtype=m.dtype)
+
+
+def tril(m, k=0):
+    """
+    Lower triangle of an array with elements above the `k`-th diagonal zeroed.
+
+    Parameters
+    ----------
+    m : array_like, shape (M, M)
+        Input array.
+    k : int, optional
+        Diagonal above which to zero elements.  `k = 0` (the default) is the
+        main diagonal, `k < 0` is below it and `k > 0` is above.
+
+    Returns
+    -------
+    tril : ndarray, shape (M, M)
+        Lower triangle of `m`, of same shape and data-type as `m`.
+
+    See Also
+    --------
+    triu : upper triangle of an array
+    """
+    if m.ndim != 2:
+        raise ValueError('input must be 2 dimensional')
+    if m.shape[0] != m.shape[1]:
+        raise NotImplementedError('input must be a square matrix')
+    if not len(set(m.chunks[0] + m.chunks[1])) == 1:
+        msg = ('All chunks must be a square matrix to perform lu decomposition. '
+               'Use .rechunk method to change the size of chunks.')
+        raise ValueError(msg)
+
+    rdim = len(m.chunks[0])
+    hdim = len(m.chunks[1])
+    chunk = m.chunks[0][0]
+
+    token = tokenize(m, k)
+    name = 'tril-' + token
+
+    dsk = {}
+    for i in range(rdim):
+        for j in range(hdim):
+            if chunk * (j - i + 1) < k:
+                dsk[(name, i, j)] = (m.name, i, j)
+            elif chunk * (j - i - 1) < k <= chunk * (j - i + 1):
+                dsk[(name, i, j)] = (np.tril, (m.name, i, j), k - (chunk * (j - i)))
+            else:
+                dsk[(name, i, j)] = (np.zeros, (m.chunks[0][i], m.chunks[1][j]))
+    dsk = sharedict.merge(m.dask, (name, dsk))
+    return Array(dsk, name, shape=m.shape, chunks=m.chunks, dtype=m.dtype)
