+++ dask/array/tests/test_array_core.py
@@ -2637,76 +2637,6 @@ def test_cumulative():
         x.cumsum(axis=-4)
 
 
-def test_eye():
-    assert_eq(da.eye(9, chunks=3), np.eye(9))
-    assert_eq(da.eye(10, chunks=3), np.eye(10))
-    assert_eq(da.eye(9, chunks=3, M=11), np.eye(9, M=11))
-    assert_eq(da.eye(11, chunks=3, M=9), np.eye(11, M=9))
-    assert_eq(da.eye(7, chunks=3, M=11), np.eye(7, M=11))
-    assert_eq(da.eye(11, chunks=3, M=7), np.eye(11, M=7))
-    assert_eq(da.eye(9, chunks=3, k=2), np.eye(9, k=2))
-    assert_eq(da.eye(9, chunks=3, k=-2), np.eye(9, k=-2))
-    assert_eq(da.eye(7, chunks=3, M=11, k=5), np.eye(7, M=11, k=5))
-    assert_eq(da.eye(11, chunks=3, M=7, k=-6), np.eye(11, M=7, k=-6))
-    assert_eq(da.eye(6, chunks=3, M=9, k=7), np.eye(6, M=9, k=7))
-    assert_eq(da.eye(12, chunks=3, M=6, k=-3), np.eye(12, M=6, k=-3))
-
-    assert_eq(da.eye(9, chunks=3, dtype=int), np.eye(9, dtype=int))
-    assert_eq(da.eye(10, chunks=3, dtype=int), np.eye(10, dtype=int))
-
-
-def test_diag():
-    v = np.arange(11)
-    assert_eq(da.diag(v), np.diag(v))
-
-    v = da.arange(11, chunks=3)
-    darr = da.diag(v)
-    nparr = np.diag(v)
-    assert_eq(darr, nparr)
-    assert sorted(da.diag(v).dask) == sorted(da.diag(v).dask)
-
-    v = v + v + 3
-    darr = da.diag(v)
-    nparr = np.diag(v)
-    assert_eq(darr, nparr)
-
-    v = da.arange(11, chunks=11)
-    darr = da.diag(v)
-    nparr = np.diag(v)
-    assert_eq(darr, nparr)
-    assert sorted(da.diag(v).dask) == sorted(da.diag(v).dask)
-
-    x = np.arange(64).reshape((8, 8))
-    assert_eq(da.diag(x), np.diag(x))
-
-    d = da.from_array(x, chunks=(4, 4))
-    assert_eq(da.diag(d), np.diag(x))
-
-
-def test_tril_triu():
-    A = np.random.randn(20, 20)
-    for chk in [5, 4]:
-        dA = da.from_array(A, (chk, chk))
-
-        assert np.allclose(da.triu(dA).compute(), np.triu(A))
-        assert np.allclose(da.tril(dA).compute(), np.tril(A))
-
-        for k in [-25, -20, -19, -15, -14, -9, -8, -6, -5, -1,
-                  1, 4, 5, 6, 8, 10, 11, 15, 16, 19, 20, 21]:
-            assert np.allclose(da.triu(dA, k).compute(), np.triu(A, k))
-            assert np.allclose(da.tril(dA, k).compute(), np.tril(A, k))
-
-
-def test_tril_triu_errors():
-    A = np.random.randint(0, 11, (10, 10, 10))
-    dA = da.from_array(A, chunks=(5, 5, 5))
-    pytest.raises(ValueError, lambda: da.triu(dA))
-
-    A = np.random.randint(0, 11, (30, 35))
-    dA = da.from_array(A, chunks=(5, 5))
-    pytest.raises(NotImplementedError, lambda: da.triu(dA))
-
-
 def test_atop_names():
     x = da.ones(5, chunks=(2,))
     y = atop(add, 'i', x, 'i', dtype=x.dtype)
